+++ src/ska_mid_cbf_mcs/fsp/fsp_corr_subarray_component_manager.py
@@ -75,6 +75,8 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
 
         self.output_link_map = [[0, 0] for _ in range(40)]
 
+        self.last_hps_scan_configuration = ""
+
     # -------------
     # Communication
     # -------------
@@ -91,6 +93,9 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
             try:
                 self._proxy_hps_fsp_corr_controller = context.DeviceProxy(
                     device_name=self._hps_fsp_corr_controller_fqdn
+                )
+                self._proxy_hps_fsp_corr_controller.set_timeout_millis(
+                    self._lrc_timeout * 1000
                 )
             except tango.DevFailed as df:
                 self.logger.error(
@@ -197,6 +202,7 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
         self.frequency_slice_id = 0
         self.scan_id = 0
         self.config_id = ""
+        self.last_hps_scan_configuration = ""
 
         # release all assigned VCC to reset to IDLE state
         self._release_vcc(self.vcc_ids.copy())
@@ -278,12 +284,11 @@ class FspCorrSubarrayComponentManager(CbfObsComponentManager):
         self._assign_vcc(configuration["corr_vcc_ids"])
 
         # Issue ConfigureScan to HPS FSP Corr controller
+
         if not self.simulation_mode:
             hps_fsp_configuration = self._build_hps_fsp_config(configuration)
+            self.last_hps_scan_configuration = hps_fsp_configuration
             try:
-                self._proxy_hps_fsp_corr_controller.set_timeout_millis(
-                    self._lrc_timeout * 1000
-                )
                 self._proxy_hps_fsp_corr_controller.ConfigureScan(
                     hps_fsp_configuration
                 )