+++ src/ska_mid_cbf_mcs/fsp/fsp_corr_subarray_device.py
@@ -47,6 +47,8 @@ class FspCorrSubarray(CbfObsDevice):
 
     HpsFspCorrControllerAddress = device_property(dtype="str")
 
+    LRCTimeout = device_property(dtype=("str"))
+
     # ----------
     # Attributes
     # ----------
@@ -105,6 +107,17 @@ class FspCorrSubarray(CbfObsDevice):
         """
         return self.component_manager.frequency_slice_id
 
+    @attribute(
+        dtype="str", doc="The last valid FSP scan configuration sent to HPS."
+    )
+    def lastHpsScanConfiguration(self: FspCorrSubarray) -> str:
+        """
+        Read the last valid FSP scan configuration of the device sent to HPS.
+
+        :return: the current last_hps_scan_configuration value
+        """
+        return self.component_manager.last_hps_scan_configuration
+
     # --------------
     # Initialization
     # --------------
@@ -158,6 +171,7 @@ class FspCorrSubarray(CbfObsDevice):
 
         return FspCorrSubarrayComponentManager(
             hps_fsp_corr_controller_fqdn=self.HpsFspCorrControllerAddress,
+            lrc_timeout=int(self.LRCTimeout),
             logger=self.logger,
             attr_change_callback=self.push_change_event,
             attr_archive_callback=self.push_archive_event,
@@ -165,6 +179,7 @@ class FspCorrSubarray(CbfObsDevice):
             communication_state_callback=self._communication_state_changed,
             obs_command_running_callback=self._obs_command_running,
             component_state_callback=self._component_state_changed,
+            admin_mode_callback=self._admin_mode_perform_action,
         )
 
     # -------------