+++ tests/test_utils.py
@@ -19,10 +19,10 @@ def device_online_and_on(
     :param device_under_test: A fixture that provides a
         :py:class:`CbfDeviceProxy` to the device under test, in a
         :py:class:`tango.test_context.DeviceTestContext`.
-    :param event_tracer: A :py:class:`TangoEventTracer` used to recieve subscribed change events from the device under test.
+    :param event_tracer: A :py:class:`TangoEventTracer` used to receive subscribed change events from the device under test.
     """
     # Set a given device to AdminMode.ONLINE and DevState.ON
-    device_under_test.simulationMode == SimulationMode.FALSE
+    device_under_test.simulationMode = SimulationMode.FALSE
     device_under_test.adminMode = AdminMode.ONLINE
 
     assert_that(event_tracer).within_timeout(